Update manage-database-space.md
Hi ! Spoke with Daniel, who also presented a suggestion for this article. I do agree that the information regarding "Best practices for a tidy and clean environment." should be presented right at the beginning of the document. Nonetheless, we have some internal documentation (Freemium Database Storage Issues) which has some useful details that can be shared publicly with the customer to help them to free some application data.